NSDMem/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.0.0=pypi_0
aiohttp=3.8.6=pypi_0
aiosignal=1.3.1=pypi_0
albumentations=1.3.1=pypi_0
altair=5.1.2=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
argparse=1.4.0=pypi_0
asttokens=2.4.1=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.1.0=pypi_0
backcall=0.2.0=pypi_0
backports-zoneinfo=0.2.1=pypi_0
blinker=1.6.3=pypi_0
ca-certificates=2023.08.22=h06a4308_0
cachetools=5.3.2=pypi_0
certifi=2023.7.22=pypi_0
charset-normalizer=3.3.1=pypi_0
click=8.1.7=pypi_0
clip=1.0=dev_0
contourpy=1.1.1=pypi_0
cycler=0.12.1=pypi_0
decorator=5.1.1=pypi_0
diffusers=0.21.4=pypi_0
docker-pycreds=0.4.0=pypi_0
einops=0.7.0=pypi_0
executing=2.0.1=pypi_0
fairscale=0.4.13=pypi_0
filelock=3.13.0=pypi_0
fonttools=4.43.1=pypi_0
frozenlist=1.4.0=pypi_0
fsspec=2023.10.0=pypi_0
ftfy=6.1.1=pypi_0
future=0.18.3=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.40=pypi_0
google-auth=2.23.3=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
grpcio=1.59.0=pypi_0
h5py=3.10.0=pypi_0
himalaya=0.4.2=pypi_0
huggingface-hub=0.17.3=pypi_0
idna=3.4=pypi_0
imageio=2.31.6=pypi_0
imageio-ffmpeg=0.4.9=pypi_0
importlib-metadata=6.8.0=pypi_0
importlib-resources=6.1.0=pypi_0
invisible-watermark=0.2.0=pypi_0
ipython=8.12.3=pypi_0
jedi=0.19.1=pypi_0
jinja2=3.1.2=pypi_0
joblib=1.3.2=pypi_0
jsonschema=4.19.1=pypi_0
jsonschema-specifications=2023.7.1=pypi_0
kiwisolver=1.4.5=pypi_0
kornia=0.7.0=pypi_0
lazy-loader=0.3=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
lightning-utilities=0.9.0=pypi_0
markdown=3.5=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.3=pypi_0
matplotlib=3.7.3=pypi_0
matplotlib-inline=0.1.6=pypi_0
mdurl=0.1.2=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.0.4=pypi_0
ncurses=6.4=h6a678d5_0
networkx=3.1=pypi_0
nibabel=5.1.0=pypi_0
nltk=3.8.1=pypi_0
numpy=1.24.4=pypi_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=8.9.2.26=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu12=2.18.1=pypi_0
nvidia-nvjitlink-cu12=12.3.52=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
oauthlib=3.2.2=pypi_0
omegaconf=2.3.0=pypi_0
opencv-python=4.8.1.78=pypi_0
opencv-python-headless=4.8.1.78=pypi_0
openssl=1.1.1w=h7f8727e_0
packaging=23.2=pypi_0
pandas=2.0.3=pypi_0
parso=0.8.3=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=10.0.1=pypi_0
pip=24.0=pypi_0
pkgutil-resolve-name=1.3.10=pypi_0
platformdirs=4.2.2=pypi_0
prompt-toolkit=3.0.39=pypi_0
protobuf=4.24.4=pypi_0
psutil=6.0.0=pypi_0
ptyprocess=0.7.0=pypi_0
pudb=2023.1=pypi_0
pure-eval=0.2.2=pypi_0
pyarrow=13.0.0=pypi_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycocotools=2.0.7=pypi_0
pydeck=0.8.1b0=pypi_0
pydeprecate=0.3.1=pypi_0
pygments=2.16.1=pypi_0
pyparsing=3.1.1=pypi_0
python=3.8.5=h7579374_1
python-dateutil=2.8.2=pypi_0
python-graphviz=0.20.3=pypi_0
pytorch-lightning=1.4.0=pypi_0
pytorchtools=0.0.2=pypi_0
pytz=2023.3.post1=pypi_0
pywavelets=1.4.1=pypi_0
pyyaml=6.0.1=pypi_0
qudida=0.0.4=pypi_0
readline=8.2=h5eee18b_0
referencing=0.30.2=pypi_0
regex=2023.10.3=pypi_0
requests=2.31.0=pypi_0
requests-oauthlib=1.3.1=pypi_0
rich=13.6.0=pypi_0
rpds-py=0.10.6=pypi_0
rsa=4.9=pypi_0
safetensors=0.4.0=pypi_0
scikit-image=0.17.2=pypi_0
scikit-learn=1.1.3=pypi_0
scipy=1.10.1=pypi_0
sentry-sdk=2.12.0=pypi_0
setproctitle=1.3.3=pypi_0
setuptools=69.5.1=pypi_0
six=1.16.0=pypi_0
smmap=5.0.1=pypi_0
sqlite=3.41.2=h5eee18b_0
stack-data=0.6.3=pypi_0
streamlit=1.28.0=pypi_0
sympy=1.12=pypi_0
taming-transformers=0.0.1=dev_0
tenacity=8.2.3=pypi_0
tensorboard=2.14.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
test-tube=0.7.5=pypi_0
textblob=0.18.0.post0=pypi_0
threadpoolctl=3.2.0=pypi_0
tifffile=2023.7.10=pypi_0
timm=0.9.8=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.14.1=pypi_0
toml=0.10.2=pypi_0
toolz=0.12.0=pypi_0
torch=1.12.1+cu116=pypi_0
torch-fidelity=0.3.0=pypi_0
torchaudio=0.12.1+cu116=pypi_0
torchmetrics=0.5.0=pypi_0
torchvision=0.13.1+cu116=pypi_0
torchviz=0.0.2=pypi_0
tornado=6.3.3=pypi_0
tqdm=4.66.1=pypi_0
traitlets=5.12.0=pypi_0
transformers=4.34.1=pypi_0
triton=2.1.0=pypi_0
typing-extensions=4.8.0=pypi_0
tzdata=2023.3=pypi_0
tzlocal=5.2=pypi_0
urllib3=2.0.7=pypi_0
urwid=2.2.3=pypi_0
urwid-readline=0.13=pypi_0
validators=0.22.0=pypi_0
wandb=0.17.5=pypi_0
watchdog=3.0.0=pypi_0
wcwidth=0.2.8=pypi_0
werkzeug=3.0.1=pypi_0
wheel=0.43.0=pypi_0
xz=5.4.2=h5eee18b_0
yarl=1.9.2=pypi_0
zipp=3.17.0=pypi_0
zlib=1.2.13=h5eee18b_0
